pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "saddle"
version = "0.1.0"
description = "Saddle: standalone team-orchestrated coding assistant runtime"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Saddle Team" }]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "fastapi>=0.111.0",
  "uvicorn>=0.30.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
]

[project.scripts]
saddle = "saddle.cli:app"

[tool.hatch.build.targets.wheel]
only-include = ["saddle"]

[tool.hatch.build.targets.wheel.sources]
"saddle" = "src/saddle"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
